Factor out some common stuff from calculator.ts (#12)
I want to copy this file for the RI calculator. To avoid too much
duplication, this factors out some common logic: the "fetch from API
and interpret errors" code, and the static footer.

It also changes the `api-path` attribute of the calculator element to
`api-host`, which should contain only the protocol and hostname, not
the path. Since the RI calculator will be fetching multiple paths, I
want to separate path from host. `api-path` is not publicly
documented, so I think it's fine to remove it.