Add "Married Filing Separately" option
## Description

Also relabel "Joint" as "Married Filing Jointly" to match.

## Test Plan

Fill in form with the new value and make sure calculation succeeds.
You can tell the new value is used by entering an income of 100k and
looking at the EV incentives: MFS should be eligible for the new-EV
credit but not used-EV, and MFJ should be eligible for both.