Get API key out of embed codebase (#16)
## Description

We're almost certainly going to open-source this codebase at some
point, so let's get the API key out of it, since it's technically a
secret.

## Test Plan

Run locally and make sure the calculator works.

Run locally with an incorrect key in `.env.local` and make sure it
results in "Authorization failed".

Vercel build to be tested as part of this PR process.

## Next Steps

After deploying this change with a new key set in Vercel, make sure
the old key isn't being used elsewhere (i.e. the first-party
calculator), then revoke it.