Non-RI experience
## Description

This allows you to enter any ZIP the backend recognizes, and removes
the map+utility selector part if the ZIP is not in RI. It got a little
complicated because of the need to accommodate changing the utility
selector without losing the outline map; see the comment on the
`tempState` property for more detail on that. There are other options
for dealing with this quirk, but this felt like the straightest path.

One thing I wrestled with is that this would be easier if the
`/utilities` response included the state; that would mean the whole
`tempState` hack wouldn't be needed. But that felt too much like
allowing the needs of this specific frontend to dictate the design of
the backend.

### Followup

I'll implement the behavior of the `state` attribute on the calculator
element "locking" the calculator to that state in a followup PR.

## Test Plan

Calculate the RI zip 02814, which has multiple utility options. Make
sure the UI progresses straight from loading spinner to fully loaded
with utility selector + incentives. RI Energy should be auto-chosen.

Choose Pascoag in the utility selector; make sure the map+selector
remain visible while the new incentives load.

Calculate the zip 02859, which also has RIE and Pascoag. The
map+selector should disappear while loading, then reappear with
Pascoag still selected; incentives should include Pascoag ones.

Calculate the zip 02116, which is in MA. Only IRA incentives should
appear, with no map+selector.

Calculate a RI zip again; make sure the RI experience reappears.