Polish form grid (#38)
## Description

- The design spec has a 2-column grid on wide and medium layouts, so
  do that.

- Pin the Calculate button to the right-hand column. This will prevent
  it from jumping around when we add the email field (#37) which
  disappears after form submission.

How did people do anything before grid and flexbox were invented??

Closes https://app.asana.com/0/1205057814651000/1205636582326127

## Test Plan

Make sure the button is on the right, with an empty space to its
left. Make sure it looks the same as before on narrow (1-column)
layout.

Add another div before the button; make sure it fills in the space to
the left of the button.