Add email field (#37)
## Description

The `type="email"` attribute on the input causes browsers to validate
that the input is a plausible-looking email address, if not empty.
However, it's not perfect; e.g. `asdf@a` is accepted, though I don't
think it's a valid email.

The behavior mimics the signup field on the existing IRA calculator,
which hides the email field immediately on submit, without waiting to
see whether the submit succeeds. There's no recovery in the event of
an error.

Whether the email has already been submitted is stored in local
storage, same as the existing IRA calculator. I'm not sure how this
will work when the embed is on a third-party site; my guess is that
it's browser- and config-dependent. Off the top of my head I think it
will work as normal in browsers that partition cross-domain
cookies/localStorage (e.g. Firefox, Brave) but not in browsers that
just block cross-domain cookies/localStorage? In any case, I doubt
we'll have the field visible on third-party sites at first, so maybe
this isn't a concern for now.

## Test Plan

- Submit the form with email empty; make sure (by watching network in
  web inspector) there's no submission to the signup endpoint and the
  calculator works as normal

- Submit the form with `asdf` as the email; make sure it gets blocked
  with an error message saying the address isn't valid

- Submit the form with `test@test.com`; this makes the endpoint return
  success but doesn't add anything to Mailchimp. Make sure the email
  field disappears.