scroll results into view when loaded (#28)
https://app.asana.com/0/1205057814651000/1205571641895785/f

Playing around with how to do this. It was surprisingly squirrely.

This approach uses the lifecycle hook in the LitElement to scroll down
following an update, which should be called after render. However I was
having a hard time selecting the at-a-glance section with
getElementById. Not sure if I was just psyching myself out. This
approach seems to work now, and it's relatively small amount of code.

Question: should this actually scroll down to the loading element, so
that the feedback from submitting the form is instant? Currently waiting
until both async tasks are complete.

Demo:


https://github.com/rewiringamerica/embed.rewiringamerica.org/assets/39635/2ca458de-f61e-475a-9e76-2e64931cd0d5

Update: scrolling to the "Incentives available to you..." section if
it's shown, "at a glance" if not.