State locking
## Description

This uses the new `location` returned in the `/utilities` response to
lock the calculator to a specific state if the HTML attribute `state`
is set.

Having this in the `/utilities` response allows for a lot of cleanup.
There's no more need for the `tempState` property to hold on to the
last-fetched location while the main task is loading.

This is based on the API PR rewiringamerica/api.rewiringamerica.org#167.

## Test Plan

### With the `state` attribute set to `RI`

- Enter 02116 (non-RI), and make sure it says "that ZIP code is not in
  Rhode Island".

- Enter 02814 and make sure you get results, with two utility
  options. Switch between the two utilities and make sure the
  map+selector stay visible while loading, and incentives for the new
  utility come up.

- Enter 02807 and make sure the map+selector disappear during loading,
  then reappear.

- Enter 02116 again and make sure the map+selector disappear, and the
  error message reappears.

### With the `state` attribute unset

Same as above, but the first and last steps should display IRA
incentives only, with no map+selector.