Handle updated /utilities response format (#41)
## Description

rewiringamerica/api.rewiringamerica.org#161 will change the format of
the `/utilities` response in a backward-incompatible way, so
temporarily make the frontend able to handle both formats.

This relies on there being no utility with the ID `utilities`.

### Followup

Once the API change is deployed to prod, I'll follow up by removing
the adaptation, and also reworking the frontend logic to (a) implement
locking to one state and (b) remove the `tempState` hack.

## Test Plan

Run this against my local API server with the change applied, and
against prod. Make sure the utility selector is populated correctly in
both cases.