Save form values in local storage (#48)
## Description

Closer to parity with the IRA calculator, this saves your form inputs
in local storage when you hit the Calculate button, and restores them
on page load if available. It made sense to implement the reset button
at the same time, so that's here too.

I wrote a "safe local storage" wrapper class that silently drops
exceptions from the local storage API. This means we'll use local
storage if available, and degrade to not storing anything if local
storage is not available.

The state management is starting to get a little iffy. The way Lit
works is like React with uncontrolled form components, meaning there's
this whole layer of state (in the form elements themselves) separate
from what the Lit component is tracking, and in this interaction
keeping them in sync is nontrivial. There are several values for each
form field all in play:

- Our hardcoded default

- The value in the HTML attribute, maybe

- The value the user enters, tracked by the DOM element

- The value in local storage

- The property in the Lit component

The outline of how it's all managed is

- The component's properties hold the values that get sent to the
  server. They get updated on two occasions:

  - When the **component is mounted**, the properties are populated
    from local storage, falling back to HTML attributes if there's
    nothing in storage, falling back to hardcoded defaults.

  - When the **form is submitted**, the properties are populated from
    form data (i.e. pulling from the state tracked by the form DOM
    elements).

- When the component **renders**, the values in the DOM nodes are set
  from the component's properties. This required a change in the
  currency input's internal state management (more state yay!), plus
  Lit's `live` directive for the zip field.

This is still workable, but it's definitely a lot to keep track of. I
don't know that there's a better solution, though; is it even a thing
to use Lit to do fully controlled form components à la React?

## Test Plan

1. Change all the form fields, don't hit Calculate. Hit reset and make
   sure all the values go back to their defaults (our hardcoded values).

2. Put values in the HTML attributes for form fields and do step 1
   again; make sure the reset puts the values back to the ones from
   the attributes.

3. Refresh the page (still without hitting Calculate) and make sure
   the form fields go back to default.

4. With values filled out, hit Calculate. Watch network tab in
   inspector to make sure the right values are getting sent. Look in
   local storage to see the values stored there.

5. Refresh the page and make sure your values are still there.

6. Hit reset, then fill out the minimum necessary to submit the form
   and hit Calculate. Make sure the values sent to the server are
   exactly what's shown in the form. Make sure local storage has no
   values in it.