Apply hover styling to Calculate button on focus
This is Tom's suggestion from the a11y review doc: style the button as
if it's being hovered when it's focused, with the darker purple
background. I think we're considering this a stopgap until we get a
specific design for focus state.

https://app.asana.com/0/1205057814651000/1205636787130046

## Test Plan

Hit Tab until the button is focused; verify that it gets a slightly
darker purple background, and no focus ring.