Foundation of localization for state calculator (#62)
## Description

This uses Lit's `localize` package, which seems to be a pretty simple
and developer-friendly solution. The expected flow is:

- Wrap your localizable strings with the `msg` function.

- Run `lit-localize extract`, which statically extracts the wrapped
  strings into an XLIFF file (an XML schema for l10n).

- Modify the XLIFF file to add translations in the target language.

- Run `lit-localize build`, which reads the translations from XLIFF
  and writes them into a Typescript file for the `msg` function to use
  at runtime.

This PR:

- Wraps all the localizable strings with `msg`. (I ignored the old
  embed for now, because it's been non-localized this whole time and
  its future is uncertain. If we want it localized, it won't be too
  much work to do that.)

- Adds config for the `lit-localize` tool.

The infrastructure to actually use translated strings at runtime will
be in a followup PR.

### Note on Internationalization Sin

Piecing together a sentence, or sentence fragments, from individually
localized fragments (as is done in `state-incentive-details.ts`) is
always a risky business, because the grammar of some languages may
require things like gender and number to agree across different
fragments. As far as I know we don't yet have that problem, so I'm OK
with punting for now.

## Test Plan

Look at the site and make sure it looks exactly the same as before.

Run `lit-localize extract` and make sure a correct-looking XLIFF file
gets generated. I haven't yet nailed down the process we're going to
use for translation so the XLIFF file isn't checked in here; I haven't
decided if it should be.

I did the `msg` tagging by literally looking at every line of every
file. It's possible I missed something; we'll find out later on after
we get actual Spanish translations in.