Rewrite README for open source (#68)
## Description

This is a from-scratch rewrite of the README, with the target audience
of an external developer or potential consumer of the API. There's
also a CONTRIBUTING.md with a guide to the codebase. Information
specific to RA's deployment of the code has been moved to our Notion
instance, linked from the "Incentives / Savings Calculator" page.

I'm going with a slightly different stance on contributions for this
repo. There's more possible work here that doesn't intersect with our
internal incentives data roadmap, so I'm saying we're tentatively open
to external PRs. I'm also saying we want feature requests; I think
this is a more sensible and likely venue for those than the incentives
API.

I'll follow up with a license file and issue templates for bug reports
and feature requests.

## Test Plan

Read it