Add Tailwind; convert AuthorityLogos as PoC
## Description

This allows for gradual conversion. I'm starting with AuthorityLogos
again because it's simple and self-contained.

I stole the colors and font sizes from `consumer`. There must be a
better way to share the design system between the two codebases, but
this works for now.

## Test Plan

Looking at the authority logos, flip back and forth between this and
prod, making sure they're pixel-identical. Check the precise text
color values to be sure.