Add helper text to email field
## Description

Most of this PR is dealing with the consequent layout changes. The
form was previously a grid with `align-items: end`; this changes it to
`start` instead, and to deal with that:

- Add space above the Calculate button in non-narrow layouts to line
  it up with the email field. The space is there even if the email
  field isn't. (The current state is short-term, because I'll be
  moving the utility field into the main form soon, which will change
  the parity of the number of fields; the space will look better once
  that's done.)

- Standardize the markup for the form field labels, and make some
  adjustments to their default Shoelace styles

- Adjust the left padding of the form fields' content to line it up
  with the helper text.

TODO:

- [ ] Spanish translation of the helper text

## Test Plan

The layout is intentionally not pixel-identical to prod, but I believe
it's closer to what's in the designs than before. Notably, the space
above the labels due to them being `inline-block` has been there the
whole time unintentionally. This also correctly implements the left
padding inside the form fields.

Look at the layout in the Cartesian product of:
(with email field | without email field) x (narrow | medium | wide)

Make sure the project selector looks OK with nothing selected and with
something selected.