set modifyValueOnWheel to false on AutoNumeric (#86)
## Description

Per https://github.com/autoNumeric/autoNumeric/issues/776 a sideways
scroll on a Mac trackpad in an AutoNumeric-enabled input field can
trigger an exception. This is benign but also the AutoNumeric behavior
to adjust the number based on mouse wheel events doesn't seem super
necessary or important. We see this exception quite a bit in Sentry for
the new calculator, e.g.
https://rewiring-america.sentry.io/issues/4717766143/?alert_rule_id=14775969&alert_type=issue&environment=vercel-production&notification_uuid=5f5dd328-5c7f-4532-9aa6-b214e8057f74&project=4506118745096192&referrer=slack
so this PR disables it.

## Test Plan

Open the preview build, select the income field, and scroll sideways
with the trackpad (scroll to the left because scrolling to the right may
trigger the browser's "back" action). The field should not change and
there should be no console errors. Compare with the same behavior in the
current production build.