Add privacy copy to top of form (#92)
## Description

Very simple. Although this is starting to bring home the idea that we
need to get more systematic about representing the RA design system in
this codebase, whether by adding Tailwind or just factoring out
variables like the colors and font sizes.

TODO:

- [x] Spanish translation

## Test Plan

Look at it in all three widths.