Put the utility selection field in the main form (#91)
## Description

The calculator's API-fetching state management is totally reworked,
and hugely simplified. There's no more two-stage fetch when the user
clicks the button, just a single `/calculator` fetch. There is a
separate fetch that runs in the background after the zip field's value
changes, but these two fetches don't interact in any way. The
utilities-fetching Task is auto-run again, since it just responds to
changes in the `zip` field.

I added the ability to put a loading spinner in `sl-select`s. It's a
bit of a hack: it's in the "prefix icon" slot, moved to the right
using `order`. Flexbox is the best.

The state outline maps are gone! RIP

Next steps:

- [x] Spanish translations. I haven't done anything there yet in case
  there are changes to the English copy.

- [x] Remove `show-email` from state-calculator.html; it's just there
  for layout demonstration

## Test Plan

- Enter 02814 for the zip and make sure two utility options load (RIE
  and PUD). Pick one and calculate; make sure that utility's HVAC
  incentives come up. Pick the other one and calculate again; make
  sure that utility's HVAC incentives come up.

  Make sure the loading spinner is shown while the utilities load.

- Enter 02814 for the zip and choose PUD for the utility; submit the
  form. Refresh the page and make sure those values are still
  populated. (I.e. that the utility selection is saved in local
  storage, and gets restored after the utilities load on refresh)

- Enter 02116 (MA) for the zip and make sure the "no utility data"
  message comes up and the menu stays disabled.