Introduce React
## Description

This is a proof of concept for the bottom-up approach: starting with
converting small self-contained components and rendering them into
individual React roots, then gradually moving up the tree to the
top-level `rewiring-america-state-calculator` component. The end state
will still be a web component; it will wrap a single React root that
renders the entire UI.

The gist is to render empty divs within `render()`, then create React
roots for them within `update()`, which is called after the changes
from `render()` are reflected in the DOM. This probably isn't great
for performance, but it's a temporary state.

The size of the production `state-calculator.js` goes from ~530K to
~704K, which isn't terrible.

## Test Plan

The rendered authority logos section is pixel-identical to prod. If
there is a flicker when it first renders, I can't see it, because it
happens scrolled out of view.