Use external icon files in icon-tab-bar
## Description

Rather than having SVG inline within code files, this uses the same
external icon files that are used for the project multiselect.

Also a drive-by fix for a incorrect ARIA labels on the pills: it was
stringifying a function rather than calling the function to get a
string. (This wouldn't happen with React/JSX!)

## Test Plan

Flipping back and forth between this state and prod, they look pixel-identical.