only request items we know about (#87)
## Description

We want to add new items to the API without breaking the calculator. The
`?items=` functionality already exists in the API, and the embed code
already knows which items we care about, so this just wires the two
together.

## Test Plan

Test with each project type in turn, especially HVAC and other projects
with multiple item types. Inspect the API request to make sure it's
setting `?items=` multiple times.

Note that the `?items=` in the API request will always be the same, no
matter which projects are set in the UI, because the multi select just
determines which projects are sorted to the top vs placed in the "other"
section.