Update to the way React elements are rendered (#99)
## Description

Working on more sophisticated components revealed a problem with the
previous approach: calling `createRoot` repeatedly with the same DOM
node.

To prevent this, we now store the React `Root` and corresponding DOM
node for each React element that gets rendered. During the Lit update
cycle, it reuses the existing stored React root if it's present and
the DOM node is still in the document.

The logic is factored out because the old embed will need to share it
soon.

Also, a drive-by fix to AuthorityLogos: the img elements needed keys.

## Test Plan

Use the state calculator, in particular alternating between
calculating for a RI zip code and then a MA zip code (the important
thing being that one state will show AuthorityLogos and one
won't). Make sure the authority logos show up, or don't show up, as
appropriate.

Make sure the console is free of any React error messages.