Convert state incentive details to React (#103)
## Description

This is relatively straightforward, as there isn't any complex
interaction.  The only snag is the increased timeout in
`waitAndScrollTo`, to wait for the targeted divs to appear in the DOM.

Also had to change the config for the Cypress workflow; becoming a
public project seems to have stopped the Cypress runner from working?