Convert the rest of the old calculator, plus footer (#104)
## Description

All of the content in both new and old calculators is React now. Lit
is still used for the top-level web components, and for the `css`
template; I think for simplicity it'll be worth getting rid of both of
those usages and dropping the Lit package. That's the next stage.

## Test Plan

This is a very simple one: there are no interactive components
involved, and no state. Use the old calculator with various
inputs. Make sure the layout is pixel-identical to prod. Especially
look out for any errant dollar signs.