More Tailwind conversions (#107)
## Description

This picks off the icon tab bar and some miscellaneous stuff that was
in the top-level calculator component.

## Test Plan

Make sure the icon tab bar (in both small and non-small
configurations) are pixel-identical to prod. Same with the separator
lines (make sure the ones in incentive cards disappear in small
layout) and the loading state.