Reduce new calculator's reliance on shared styles (#112)
## Description

This is the last of the Tailwind transition, for now. It breaks the
new calculator's dependency on styles shared with the old calculator;
now the only shared parts left are:

- Styles on `:host` (i.e. the web component itself)
- Replacing the default styling of `<input>` elements
- The tooltip and select components

To do this:

- All the h1, h2, and p elements now need explicit Tailwind styles for
  size and line height

- The `calculator` class on the root div needs to be replaced

- The `calculator-footer` class on the footer needs to have Tailwind
  styles in addition to the semantic class name, because it's shared
  by both calculators and the old one doesn't use Tailwind

## Test Plan

Make sure this and prod are pixel-identical at all three sizes,
including the no-results state and the authority logos.

Audit all the `<h1>`, `<h2>` and `<p>` elements in the new calculator
to make sure they have Tailwind styles that set font-size and
line-height.