Fix compliance with the "state" attribute
## Description

The "state" attribute, if set, is supposed to cause the calculator to
display an error message (and no incentives) if the user enters a zip
code outside of that state.

This stealthily broke when we moved the utility selector into the main
form. Previously, the top-level error state was displayed when
_either_ the utilities fetch or the main fetch had errored, and the
utilities fetch errored if the state was wrong. After the move, a
failed utilities fetch didn't block the main fetch from happening, and
the main fetch did not check the state attribute.

This adds that check, and also adds a Cypress test because this
restricted-state configuration is kind of obscure and easy to
accidentally break.

## Test Plan

Manually add the `state="RI"` attribute and calculate for a MA zip
code; make sure the error is shown and no incentives are shown.

New Cypress test passes; watch the video to confirm it's testing the
intended thing.