Reimplement Spinner component without Shoelace
## Description

This mimics the MUI `CircularProgress` that we use on the consumer
site. It looks different from the existing Shoelace spinner (which has
an always-visible circular track, and uses different colors), but that
was just a convenient off-the-shelf thing, not a considered design
choice.

## Test Plan

Look at it in its two contexts (within the utility selector, and while
loading incentive results). Helps to turn on throttling in web
inspector so it's there for longer.